All inv src loops do name replacement, workflow docs
This changes the inventory source resource to always replace the
name field with a querystring for "starts with" logic,
this enables the direct once-through use of such resources
by the workflow schema command, combined with other commands.

Additionallly, an example is added to demonstrate this type of
use, which also leverages the parallel work for workflow
schema overwriting.